Heavy speedup of runDataFrameFinal
RDataFrames are fast, but they need to be used correctly. In
particular one has to pay attention to not access too early
the variables returned by the lazy operations.
This commit reorganizes runDataFrameFinal so that all filters,
histograms and snapshots are defined before the event loop is
triggered. That way all can be filled in one pass, instead of
n_histos x n_cuts passes before.

On the example mH-recoil/mumu, running time drops from 6min30
to 40sec.