Update of vertexing code: move to the code from Franco B that is in Delphes.
So far: the vertexing code from Franco (an old version of it) had been copied/pasted into FCCAnalyses, in VertexFitterSimple
(that was dumb of course, but easy, e.g. no need to figure out how to compile the FCCAnalyses code using delphes as a library).
Now Delphes is linked properly, and the Vertexing code of FCCAnalyses has been purged from all the code that had been copied.

Moreover: the vertexing code of Franco is now an updated version (see his presentation at the Physics Performance meeting of July 2022).

The FCCAnalyses method that retrieves the primary tracks has been updated - the old code would still work, but is very unefficient,
with plenty of the (new) VertexFit objects created in a recursive method. Performance recovered with this update.

Updated the vertexing examples to comply with the updated primary tracks retrieval.