Fix in SmearObjects::SmearedTracks
For a small fraction of the tracks ( ~ 5e-6 ), the covariance matrix is
not definite positive, with a determinant very close to zero but negative.
This is probably a numerical precision issue, because edm4hep stores the covMats as an
array of floats and not doubles. In those pathological cases, the method that produces
a new smeared track would crash -> return the original track instead.